% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dumDEL.R
\name{dumDEL}
\alias{dumDEL}
\title{dumDEL}
\usage{
dumDEL(x)
}
\arguments{
\item{x}{The df as input (can use lapply because it's just x)}
}
\value{
transformed data.frame
}
\description{
takes Historical() output and groups by Trues/Falses
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 dumDEL(LGP6122)
 }
}
}
\seealso{
\code{\link[stringr]{str_replace}},\code{\link[stringr]{str_remove}}
}
