% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historicalVI.R
\name{historicalVI}
\alias{historicalVI}
\title{historicalVI}
\usage{
historicalVI(file)
}
\arguments{
\item{file}{csv filepath for data to be pulled}
}
\value{
results for a given date of csv data
}
\description{
take csv file path and gets all needed data for analysis
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[here]{here}}
\code{\link[stringr]{str_sub}},\code{\link[stringr]{str_split}},\code{\link[stringr]{str_remove}},\code{\link[stringr]{str_locate}},\code{\link[stringr]{str_replace}}
\code{\link[zoo]{yearmon}}
\code{\link[utils]{read.table}}
\code{\link[quantmod]{getSymbols}},\code{\link[quantmod]{OHLC.Transformations}}
\code{\link[purrr]{map}},\code{\link[purrr]{reduce}},\code{\link[purrr]{reexports}}
\code{\link[data.table]{transpose}}
}
