% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MerlinQuantV1-package.R
\docType{package}
\name{MerlinQuantV1-package}
\alias{MerlinQuantV1}
\alias{MerlinQuantV1-package}
\title{MerlinQuantV1: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
