% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validSymbols}
\alias{validSymbols}
\title{Valid Symbols Dataset}
\format{An object of class \code{data.frame} with 8889 rows and 1 columns.}
\usage{
validSymbols
}
\description{
Contains every valid stock symbol we can successfully pull price data on
}
\keyword{datasets}
